Change comparison against reference and commit sequences build from Marzluff.
* scripts/reference_comparison.pl: change the script that performs comparison
against a reference. The idea is that this can be easily changed to perform
comparison against any reference data (maybe even data generated using the
same method after a few years). So off it goes the Marzluff data which is now
in data/. The script will now accept the path for a directory with new
reference data so we can compare against anything.
* scripts/MyLib.pm: accept new option "reference".
* data/reference-Marzluff_2002/README.md: new file to explain what is this
directory, how the data here was generated, and some issues we found.
* data/reference-Marzluff_2002/build_marzluff_data_2002.pl: script used to
convert the data found in the Marzluff paper into something that resembled our
generated data to help standardize comparison against a reference.
* data/reference-Marzluff_2002/data.csv: file with `database' for genes
described in Marzluff paper.
* data/reference-Marzluff_2002/proteins/*: sequences for each protein described
in Marzluff paper.