Revert "Replace our MyLib::pretty_length with Number::Format::format_bytes." This reverts commit 8f7c4a2784f6f720bbf8e2e64017048fc63f25ee.
The Number::Format module did not allow to control output as well.  We want
to have control on the space between the number and unit.  That module is
very much focused on disk space units, so there really wasn't much point
on adding complexity to it.  It also only uses a smaller number of SI prefixes.