Drop date of sequences from the list of anomalies (start new unused HistoneCatalogue.pm).
The date when the sequences were collected is important but that's
hardly unique to the list of anomalies identified.  So drop it from
the header of that table.  We will have to think of a place for it
(probably in the materials and methods of the manuscript).

Also, finally create an HistoneCatalogue package to start collecting
everything in a way that will truly make this code easy to reuse.