Add SCons tool for perl5 and adjust scripts to use library.
Moved our perl modules into lib-perl5 (we probably won't but maybe in
the future we will have a lib-perl6 or lib-mumble).

Add a perl 5 scanner as a scons tool so we don't have to mark our
dependencies manually.  This tool also handles PERL5LIB construction
environment so we can easily set @INC easily from SCons.

Until now, all the perl scripts were looking for our own MyLib.pm in the
scripts directory (via FindBin).  We change this to make use of a
construction environment.  It's not looking very good at the moment but
the plan is to move more things to the HistoneCatalogue.pm, and reduce
the scripts in size complexity.

Update list of perl modules and latex package dependencies.

Do not run the configure checks if SCons has only been asked for the
help text.

Rename publication target to manuscript.