extract sequences: add organism build option.
Move the call to bp_genbank_ref_extractor into the build system.  No need
to have a separate script for this only.  This makes it easier to have build
options control the options.  At the moment for the organism only, but later
also the search.

Part of making things more modular.

Create a new NoShellCommand builder to avoid shell interpretations
and messing up arguments.