Drop say_latex_newcommand() in favor of 'say latex_newcommand()'.
The previous function required a file handle to which it called say().
This made it needlessly complicated if we only wanted to prit to STDOUT.
We do not yet, but we will want soon as we start to redirect STDOUT of
scripts into the target file from SCons.

Adjust HistoneCatalogue.t accordingly.