Add tex file to perform FP computations outside the main document (issue #12).
Rather than doing complex computations in LaTeX using the fp package,
we now use a computations.tex that gets included at the preamble
instead of in the main text (it disrupts flow of the text, specially
if it's a multiline computation).

Now the document environment should only be used for formatting,
which includes FPround.

We could also make those computations in the scripts but sometimes
it's jut complicated or doesn't really fit there.