Move creation of csv files with set of histone genes into separate target.
The csv files for h1, core canonical, and linker histones are not used
anywhere for the manuscript.  So we move it to a separate target that is
not built by default (new target is aliased to 'csv').

Also move that part from the extract_sequeces script (which is being removed)
into a separate script create_histone_csv that makes use of our new
HistoneSequencesDB object.

Also fixes one issue with the csv files generated.  The header had one extra
column that was never used, the ensembl id.