Remove old, useless, and buggy MyLib::make_csv_catalogue.
This was creating the files `results/table-histone_catalogue.csv` and
`results/table-variant_catalogue.csv` which matched the tables from the
manuscript.  However, these are not used anywhere in the manuscript, and
there is already a create_histone_csv.pl script that creates csv with
similar purposes within `results/csv` (SCons target is `csv`).

Also, the removed code had a bug where it was missing one of the headers
and the new CSV files have more information.