Automatically choose best SI prefix in LaTeX.
We want all the formatting to be done in LaTeX.  This replaces the perl
code that automaticaly choose the best SI prefix for bp length.  This
code has the advantage that writing the units is all done in latex and
we can reuse it for non distances.  Also, now the latex commands have
the actual length, no rounding performed yet.

This indirectly fixes issue #14 (stray comma due to an escaping issue in
the perl code).