Compute histone pairwise dN (Ka), dS (Ks), and dN/dS (Ka/Ks) (issue #6)
Perform computations using PAML's codeml, following the strategy
by bioperl's bp_pairwise_kaks (dN/dS is also know has Ka/Ks).

Add new reference, fix table to use the computed (instead of the
ones values by Andrew computed via SNAP).  We are not using SNAP
because it's license is not free (it is gratis for non-commercial
use only), the code available for download does not match the one
online, and is just easier and simpler to use something already
packaged in Bioperl (on which we already depend on).