Create HistoneCatalogue::say_histone_catalogue to replace MyLib::make_tex_catalogue.
The new function makes us the HistoneSequencesDB which has a better design, is
tested, and works faster (even as a one liner).  Make use of it for the canonical
core table.  Plan is to have it replace the variant table next.

MyLib::make_tex_catalogue also has a bug when dealing with non-coding transcripts
(which appears when dealing with the variant histones -- see issue #21), which we
will fix on HistoneCatalogue::say_histone_catalogue.

Add tests and new dependency on Test::Output.