Remove text references to SNAP and CUSP (EMBOSS).
Remove text and reference about SNAP version.  We do not use SNAP for
anything (see 9a3ad8b0 for what we really do, and issue #6)

Same for CUSP from EMBOSS.  It is not being used for anything.  Andrew
seems keen on using it for codon usage but that will not happen. I'll
probably just use Bio::Tools::SeqStats and Bio::CodonUsage::Table.
See issue #5 for details..