Do not include date of build (very misleading) and other fixes on materials section.
bp_genbank_ref_extractor does not make the core of thr processing.  It
pretty much only downloads the data.

Do not include date of build.  It is already at the front of the
document and can be misleading since it may not match the date of
the sequences (which is what really matters).