Clear up in tex variables and code that we only deal with core histones (issue #22)
Throughout the text, we mention histones when we actually mean core histones.
This is also reflected in the code with variable named histones when they actually
refer to core histones only.  So clear that up on the code.  Next we need to clear
that up on the manuscript text.

This includes counting core histones only, and removing them from the histone
tables.  This may have the problem in the future that H1 wil not be counted
when measuring the length of HIST1.

And this is all because we want to be focused on the core histones but also
because we have the "undocumented" stand that H1 should not really be a histone.