We do not want to handle H1 so don't mention numbers about it.
We have decided that we will not handle H1 (that H1 should not even
be an histone).  We still mention it.  However, when it comes to actual
numbers we need to automate those numbers.  But we choose to not have
those numbers since we do not handle H1.  So remove specifics about them
that can easily get out of date.