protein_stats: rewrite for HistoneSequencesDB and issue #17
This rewrite to replace MyLib with HistoneSequencesDB accounts for
better reproducibility by counting all proteins, and not only the
first of each gene (issue #17).  Also simplifies usage, it only prints
values to stdout which are redirected to the file by SCons.

It also prints the values with whole precision so that rounding with
2 decimal cases is now performed in LaTeX.

Also, do not compute the lowest and highest ratio MaxCoreArgLysRatio
and MinCoreArgLysRatio since they are not used anywhere.