WebLogo: fix when input and output file are the same.
We were reading whole file in memory and then print out as we modify it
so it works well even when input and output are the same.
However, if something goes rong during the processing. We destroy the
input file.  This first creates a temporary file and then moves it.