sanity checks: refer to protein ID after identifying a non-annotated stem-loop.
We are dealing with identifying anomalies and one of them is an histone gene
with multiple transcripts.  If more than one transcript was missing the stem
loop annotation, the generated text would be the same for all (since it only
refers to gene symbol) which looked like a bug.  So also refer the protein
id.