Use tabularx to automatically set width of column describing isoforms.
We were using eqparbox with \newsavebox to automatically compute the
size of the column describing the isoforms differences. This has the
issue that all tables were using the same size and requires configuration
on both the preamble (were F was defined), and on the code (where F was
hardcoded).  This solution makes the code self-contained.

Also, set the text to left aligned (ragged) instead of justified.