HistoneSequencesDB: coding genes with non-coding transcripts first on csv.
The Gene constructor fails when building a coding gene without a protein
product.  However, it is possible that on the csv file, the non-coding
transcripts appear first (at the moment this happens randomly for H2FJ).
Our previous work for this failed when the first product appearing was
non-coding. Fix this and add more tests.