Set manuscript to oneside since that makes more sense for our readers.
Most people will read this on a monitor.  Also, even if printed, they
just turn the page over, it is never displayed side by side with the
next page.  Even when printed both sides of the paper, two pages are
never side-by-side which is the point of the different margins for
twoside.

Also remove 10pt since that's the default and we don't really have a
preference.