Create small report as the catalogue as default target (issue #9).
The manuscript, previously named histone-catalogue.pdf gets renamed
to manuscript.pdf.  A new document, catalogue.pdf, on teh style of
a short catalogue with the most important tables and figures is added
and set as default target.  This makes more sense since we do not have
a manuscript for all organisms but we should have for the catalogue.

Also, move the histone count table into separate file so it can be
shared between the catalogue and the human manuscript.