Split figure for histone nomenclature in 3 subfigures.
One comment received was that each one of the subfigures explaining
the nomenclature of histone should have a title.  Do this by using
subbottom (maybe we could use subtop?).

In addition, replace the svg figure with LaTeX code that simply uses
colorbox.  The alternative would be to split the svg file into 3
files but then keeping the font size between them the same becomes
tricky (the width of their text is different).

This drops our dependency on inkscape which is nice.

Replace svg figure for nomenclature with LaTeX code and split