HistoneSequencesDB: fix bug when first row of csv data is a gene with multiple products.
We were using a list to collect all genes.  We checked if the gene had
already been found by iterating the list and looking for that UID and
returning an index. If that gene was the first element, then index would
be zero, and evaluate as true.  Replace list with an hash and then make
a list of its values.

Also add a test for this corner case.