Be careful with case of histone type and gene symbol to support non-humans.
In human, all histone gene symbols are upper case but not in other
organisms.  One way to handle this is to make all comparisons case
insensitive which is easy to forget (we already have some like that),
another is to up case everything from the start which is what MyLib
was doing but then we lose the correct case when creating the output
tables.

We do need to upcase the histone type when guessing from the gene
symbol.