HistoneSequencesDB: warn instead of error when failing to create Gene object.
With the human genome this always seem to work but with less curated
genomes this causes issues.  Current issue is gene UID 625328 (H3f3c
a mice H3 variant) which is listed as protein coding but does not have
any reference protein or transcript (only predicted model).

This change means that the build will complete with incomplete data
but in a way it is correct since it maches the current annotations
which we interpret as the current knowledge.  We need to look at the
stderr of the build to make sure we don't miss anything since this
issue won't even show up on the table of anomalies.