Major rewrite of README and scons help text.
Both README and scons help text had out of date list of targets.  It
was also not clear that the list of dependencies on README was not
complete and that the scons help text should be used instead.  Fix
that.

Nicer introduction text with a better explanation what is goin on and
talking about builds.  Move more build instructions nearer the top
with description of most important targets and options and their
gotchas.  Also add examples.

Separate perl dependencies into needed for analysis and those needed
for the test suite only, and only configure check the ones needed by
the test suite when it is going to be run.