Restructure data/reference- to support multiple organisms (issue #1)
The original plan was to have each organism in a git branch.  Now, we
want to have a single source directory that can support any organism
by using build options and maybe multiple build directories (VPATH
builds).  So rename reference-Marzluff_2002 to reference-homo-sapiens
so that: 1) we can have the mus musculus reference (also from Marzluff
2002) on the same directory, and 2) deduce at configure time the path
for the reference data directory.

Remove the reference values from the preamble and include them only
for the manuscript.  And check at configure time if we have reference
data for this organism and skip it if not.