Replace ctabular with xtabular from xtab package.
While ctabular did the job fine, it is part of the memoir package only.
When we submit for a journal we will need something in a package so it's
better to sort it in master than only on that journal branch since change
here involves modifying the perl code and its tests.