elife.cls: reduce margin and skips in featurebox float (issue #37)
The featurebox environment is a float.  This causes problems if its content
is longer than one page.  We could replicate the environment with afterpage
and captionof.  However, we are only a few millimeters over one page so
instead we reduce the skip between caption and frame content and the top
frame margin.