Manage references to supplementary tables (issue #38 and issue #39)
The eLife template changes how figure and tables behave in the appendix.
But this is meant to have multiple appendixes, each appendix with multiple
figures or tables which is not really our case.

So declare our own float environment.  Couldn't figure out how to
automatically prepend a S to the ref as there is in the memoir document
class, so make a new supref command instead.

By not using floats at all, only captionof, also make sure long tables do
not break in half with floats in between.