Name all supplementary data as "tables" in the text.
The supplementary data all have the same counter and numbering "SX"
and there as no distinction between figure and table. It's weird that
S1-S3 are tables, S4-S7 are figures, and S8 onwards are tables again.
So name them all tables in text which is the only thing supported on
our upcoming elife submission.